requirements.txt
sanic==0.8.0
python-keycloak
